% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{adjustSpeed}
\alias{adjustSpeed}
\title{Funciton for adjusting speed, e.g., to speed up the survey.}
\usage{
adjustSpeed(x, fact = 1)
}
\arguments{
\item{x}{The output from \code{\link{biomass2tsb}} or \code{\link{runPelfoss}}.}

\item{fact}{A value below which NASC values are omitted when plotted as black dots with size proportional to NASC^NASCexp.}
}
\description{
Funciton for adjusting speed, e.g., to speed up the survey.
}
