% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{biomass2nasc}
\alias{biomass2nasc}
\title{Funciton for converting from a biomass density to nautical area scattering coefficient NASC (linear values, not dB).}
\usage{
biomass2nasc(Wg, LcmOne, a, b, m = 20, TS0 = -71.9)
}
\arguments{
\item{Wg}{The biomass area density in units of g/m^2 (gram per square meter in the horizontal plane).}

\item{a, b}{Parameters of the length-weight relationship Wg = a * Lcm^b, where Wg is the weight in grams and Lcm is the length in cm. Typical values are e.g., a = 0.01 and b = 3.}

\item{m, TS0}{The parameters of the target strength-length relationship TS = m * log10(Lcm) + TS0, typically m = 20 and TS0 = -71.9 (herring, from Foote, K. G. 1987. Fish target strengths for use in echo integrator surveys. Journal of the Acoustical Society of America, 82: 981– 987.)}
}
\value{

}
\description{
Funciton for converting from a biomass density to nautical area scattering coefficient NASC (linear values, not dB).
}
