% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{getFisheryWeights}
\alias{getFisheryWeights}
\title{Function for getting weights to apply to the strata based on the landings.}
\usage{
getFisheryWeights(fisheryFile, stratumFile, strata, startdate, numdays)
}
\arguments{
\item{fisheryFile}{The path to the csv file containing the catches from the fishery. The file must contain the columns DATE, in the format "yyyy-mm-dd", VEKT, and LON and LAT in decimal degrees.}

\item{stratumFile}{A file with strata definitions readable by \code{\link[Rstox]{readStrataPolygons}}.}

\item{strata}{a vector of names of the strata to include.}

\item{startdate}{A string of the start date of the survey, given as "d/m"}

\item{numdays}{An integer giving the number of days to use for the survey.}
}
\description{
Function for getting weights to apply to the strata based on the landings.
}
