% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{getRunPar}
\alias{getRunPar}
\title{Add the parameters used in the PELFOSS simulation, stored in the list of reports 'x'.}
\usage{
getRunPar(x)
}
\arguments{
\item{x}{List of reports as returned from \code{getAllTSB}.}
}
\description{
Add the parameters used in the PELFOSS simulation, stored in the list of reports 'x'.
}
