% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{getTotalBiomass}
\alias{getTotalBiomass}
\title{Interpolate NORWECOM biomass onto log distances.}
\usage{
getTotalBiomass(superInd, stratumPolygons, strataInd = NULL, days = NULL,
  type = c("biomass", "length"), lonlatFilter = NULL)
}
\arguments{
\item{x}{The input NetCDF4 file.}
}
\value{

}
\description{
Interpolate NORWECOM biomass onto log distances.
}
