% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{interpolateBiomassToTransects}
\alias{interpolateBiomassToTransects}
\title{Interpolate NORWECOM biomass onto log distances.}
\usage{
interpolateBiomassToTransects(biomass, transect, centroid, dayvec,
  margin = 0.05, proj = "aeqd", units = "kmi", x_0 = 0, y_0 = 0,
  ellps = "WGS84", datum = "WGS84")
}
\arguments{
\item{x}{The input NetCDF4 file.}
}
\value{

}
\description{
Interpolate NORWECOM biomass onto log distances.
}
