% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{pelfoss-package}
\alias{pelfoss}
\alias{pelfoss-package}
\title{This package provides code for simulating acoustic-trawl survey estimates from spatial biomass data, developed for the PELFOSS project (Observational system simulator for pelagic fish stocks).}
\description{
The function \code{runPelfoss} is the main funciton in this package, which requres a certain folder structure to be in place. The skeleton of this folder structure can be created using the function \code{createPelfossSkeleton} and contains the following folders: "data", "fishery", "reports" and "stratum". During execution additional folders "all_output", "all_plot", "project" and "XMLfiles" are automatically created.
}
\details{
The folder "data" is assumed to have the following sub-folders: SPECIES / RESOLUTION / YEAR, where SPECIES can be named arbitrarily (e.g., Herring or Mackerel); RESOLUTION is assumed to be in the form "res_" folowed by a numeric and then the string "km", e.g., "res_4km" (the resolution is always given in kilometers, as opposed to the units used for Cartesian positions which is nautical miles by default); and YEAR in the form "year_" followed by the year, e.g., "year_2010". Each sub-folder is assumed to hold NetCDF4 files two files, one for the biomass in grap per square meter, and one for the superindividuals used for generating the biomass file. This is assuming the ecological model used to generate the files contains superindividuals which move through th model region. The time resolution of the ecological model output is assumed to the days. These files are required to start with the string "biomass_" and "superind_", respectively, for the biomass file and the superindividual file. The biomass file must contain the following variables:

The folder "fishery" needs sub-folders for the species. The folder "reports" has no requred sub-folders. The folder "stratum" needs to have one sub-folder per cruise to be simulated, each containing one or several files defining the stratum polygons as readable by \code{\link{readStrataPolygons}}. The names of the cruises needs to start with the secies names used in the folder "data".
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Sea2Data/pelfoss/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@imr.no}

}
