% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{pelfoss}
\alias{pelfoss}
\title{This package provides code for simulating acoustic-trawl survey estimates from spatial biomass data, developed for the PELFOSS project (Observational system simulator for pelagic fish stocks).}
\usage{
pelfoss()
}
\description{
This package provides code for simulating acoustic-trawl survey estimates from spatial biomass data, developed for the PELFOSS project (Observational system simulator for pelagic fish stocks).
}
