% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{pelfossDefaults}
\alias{pelfossDefaults}
\alias{getPelfossSkeleton}
\alias{getPelfossPaths}
\alias{createPelfossSkeleton}
\alias{fitLengthWeightSuperind}
\alias{fitLengthWeight}
\alias{plotLengthWeight}
\title{Utility functions of the pelfoss package.}
\usage{
pelfossDefaults()

getPelfossSkeleton(dir, run = "Run1")

getPelfossPaths(dirs, run = "Run1", survey = "Herring_IESNS",
  year = 2010, res = 4, reversed = FALSE, dateshift = 0,
  seed = 0)

createPelfossSkeleton(dir, run = "Run1")

fitLengthWeightSuperind(superindDay, superind, count = "inumb",
  superindFilter = NULL, stratumPolygons = NULL, a = NULL,
  b = NULL, transectsXyOfDay = NULL, insideRange = NULL)

fitLengthWeight(L, W = NULL, a = NULL, b = NULL, count = NULL,
  filter = NULL)

plotLengthWeight(x, cols = c(1, 2), useJitter = FALSE, ...)
}
\arguments{
\item{dirs}{A list of paths to directories as returned from \code{link{getPelfossSkeleton}}.}

\item{survey}{The survey name.}

\item{year}{The year.}

\item{res}{The resoludion.}

\item{reversed}{Logical: If TRUE reverse the direction of the survey as generated by \code{\link{surveyPlanner}}.}

\item{dateshift}{The number of days to shift the timing of the survey by.}

\item{seed}{The seed used in \code{\link{surveyPlanner}}.}
}
\description{
\code{pelfossDefaults} Defines default values which can be overridden by the \code{...} option in \code{runPelfoss}. \cr \cr
\code{getPelfossPaths} Given a root directory of a pelfoss folder structure (see ?pelfoss) and a survey name, this function returns the paths to biomass and superind NetCDF4-files, stratum polygons, fishery data files (any data format) of the species as interpreted from the survey name. . \cr \cr
}
\keyword{internal}
