% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{plotPelfossReports}
\alias{plotPelfossReports}
\title{Function for plotting the PELFOSS report form one PELFOSS run.}
\usage{
plotPelfossReports(
  run,
  dir,
  var = c("TSB_5percentByInsideAll", "TSB_meanByInsideAll", "TSB_95percentByInsideAll"),
  seed = NULL,
  ylim = NULL,
  subdir = TRUE,
  case_adj = c(0.15, 0.95),
  add.cv = FALSE,
  width = NA,
  height = NA,
  adds = list(),
  report = NULL,
  face = NULL
)
}
\arguments{
\item{run}{The name of the simulation run (which is three timings (-30, 0, +30 days), two directions (normal, reversed), and two resolutions (low, high), in total 12 individual runs).}

\item{dir}{The directory of the PELFOSS folder structure (see \code{\link{getPelfossSkeleton}}).}

\item{seed}{A single integer, or a list of seeds used in the funciton, including the following seeds: ('transect') for drawing transects using surveyPlanner(), ('trawl') for drawing trawl stations along the transects with probability as a funciton of the NASC (see \code{probfact}), and ('bootstrap') for getting the final estimate using runBootstrap() (see \code{nboot}).}

\item{ylim}{The ylim of the plot, useful for setting the same ylim on several related plots.}

\item{subdir}{The folder in which to put the plots in the "reports" folder of \code{dir}. If TRUE this is interpreted as "allPlots".}

\item{case_adj}{A two element vector giving the adjustment of the labels of each survey (A, B, ...).}

\item{by}{A string specifying which variable to plot. The options are "Inside" (only inside the strata with positive NASC), InsideAll (inside the strata used in the survey) and "Total" (all possible strata in the strata system).}
}
\description{
Function for plotting the PELFOSS report form one PELFOSS run.
}
