% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{readAllPelfossReports}
\alias{readAllPelfossReports}
\title{Function for reading all reports of one PELFOSS run.}
\usage{
readAllPelfossReports(run, dir, seed = NULL)
}
\arguments{
\item{run}{The name of the simulation run (which is three timings (-30, 0, +30 days), two directions (normal, reversed), and two resolutions (low, high), in total 12 individual runs).}

\item{dir}{The directory of the PELFOSS folder structure (see \code{\link{getPelfossSkeleton}}).}

\item{seed}{A single integer, or a list of seeds used in the funciton, including the following seeds: ('transect') for drawing transects using surveyPlanner(), ('trawl') for drawing trawl stations along the transects with probability as a funciton of the NASC (see \code{probfact}), and ('bootstrap') for getting the final estimate using runBootstrap() (see \code{nboot}).}
}
\description{
Function for reading all reports of one PELFOSS run.
}
