% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{readNcVarsAndAttrs}
\alias{readNcVarsAndAttrs}
\alias{readNorwecomBiomass}
\alias{readNorwecomSuperind}
\alias{getNorwecomSuperindLonLat}
\alias{projLonLat2XY}
\alias{interpolateBiomassToTransects}
\title{Read a NORWECOM file.}
\usage{
readNcVarsAndAttrs(ncfile, vars, rename = NULL)

readNorwecomBiomass(ncfile, centroid, vars = c("Biom", "Long", "Latt"),
  rename = list(Biom = "biomass", Long = "longitude", Latt = "latitude"))

readNorwecomSuperind(ncfile, centroid, vars = c("xpos", "ypos", "Long",
  "Latt", "female", "age", "length", "pweight", "inumb"),
  rename = list(pweight = "weight"), removeIntitialNA = TRUE,
  maxValue = 1e+10)

getNorwecomSuperindLonLat(superind)

projLonLat2XY(x, centroid, proj = "aeqd", units = "kmi", x_0 = 0,
  y_0 = 0, ellps = "WGS84", datum = "WGS84")

interpolateBiomassToTransects(biomass, superind, transects, centroid, dayvec,
  lonlatmargin = 0.05, superindFilter = NULL, m = 20, TS0 = -71.9)
}
\arguments{
\item{ncfile}{The input NetCDF4 file with a (possibly irregular) grid of biomass in gram per square meters, and geographical positions (see the input \code{vars}).}

\item{vars}{A string vector of the variables to read from the \code{ncfile}.}

\item{rename}{A list of variableNameInNCfile = newVariableName, renaming the variables read from the \code{ncfile} to the variables required by the function \code{biomass2tsb} ("Biom", "Long", "Latt", for output from \code{readNorwecomBiomass}, and "gridLongInd", "gridLattInd", "gridLong", "gridLatt", "age", "inumb", "length", "sweight", "pweight", for output from \code{readNorwecomSuperind}).}

\item{centroid}{The geographical position in decimal degrees (longitude, latitude) used as centoid in the conversion to Cartesian positions.}

\item{removeIntitialNA}{Logical: If TRUE, remove superindividuals which are initially NA (thus born within the time period of the file). This}

\item{maxValue}{The value above which variables are identified as NA.}

\item{x}{The output from \code{readNcVarsAndAttrs}.}

\item{proj, units, x_0, y_0, ellps, datum}{The proj4 parameters used in the projection from geographical to Cartesian coordinates centered at \code{centroid}.}

\item{biomass, superind}{The biomass and superind data read from \code{readNcVarsAndAttrs} (used in \code{readNorwecomBiomass} and \code{readNorwecomSuperind}).}

\item{transects}{The output from \code{\link{surveyPlanner}}.}

\item{dayvec}{A vector of the days in the biomass data at which the log distances in \code{transects} are timed.}

\item{lonlatmargin}{The margin to expand the ranges of longitude and latitude when interpolating each stratum. A large value demands more CPU time and memory.}

\item{superindFilter}{An optional function acting on the superind data, such as longitude or latitude, e.g., superindFilter <- function(x) x$longitude < 20.}

\item{m, TS0}{The parameters of the target strength-length relationship TS = m * log10(Lcm) + TS0, typically m = 20 and TS0 = -71.9 (herring, from Foote, K. G. 1987. Fish target strengths for use in echo integrator surveys. Journal of the Acoustical Society of America, 82: 981 - 987.)}
}
\description{
\code{readNorwecomBiomass} Reads a NORWECOM biomass file, with biomass on a irregular grid Long, Latt. \cr \cr
\code{readNorwecomSuperind} Reads a NORWECOM super individual file. \cr \cr
\code{readNcVarsAndAttrs} Extracts variables with long names as attributes. \cr \cr
\code{getBiomassXY} Converts to Cartesian coordinates for the biomass data. \cr \cr
\code{getSuperindXY} Converts to Cartesian coordinates for the superindividual data. \cr \cr
\code{interpolateBiomassToTransects} Interpolate NORWECOM biomass onto log distances. \cr \cr
}
