% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{readNORWECOMbiomass}
\alias{readNORWECOMbiomass}
\alias{readNORWECOMsuperind}
\alias{readNcVarsAndAttrs}
\alias{addTimeHoursFrom1950}
\alias{getBiomassXY}
\alias{getSuperIndXY}
\title{Read a NORWECOM file.}
\usage{
readNORWECOMbiomass(ncfile, centroid, vars = c("MACbiom", "Long", "Latt"),
  rename = list(MACbiom = "Biom"))

readNORWECOMsuperind(ncfile, centroid, vars = c("xpos", "ypos", "Long",
  "Latt", "female", "age", "inumb", "length", "sweight", "pweight"),
  rename = list(xpos = "gridLongInd", ypos = "gridLattInd", Long = "gridLong",
  Latt = "gridLatt"), removeIntitialNA = TRUE)

readNcVarsAndAttrs(ncfile, vars, rename = NULL)

addTimeHoursFrom1950(x, timedim = "time")

getBiomassXY(biomass, centroid, proj = "aeqd", units = "kmi", x_0 = 0,
  y_0 = 0, ellps = "WGS84", datum = "WGS84")

getSuperIndXY(superInd, centroid, proj = "aeqd", units = "kmi", x_0 = 0,
  y_0 = 0, ellps = "WGS84", datum = "WGS84")
}
\arguments{
\item{x}{The input NetCDF4 file.}
}
\value{

}
\description{
\code{readNORWECOMbiomass} reads a NORWECOM biomass file, with biomass on a irregular grid Long, Latt.
\code{readNORWECOMsuperind} reads a NORWECOM super individual file.
\code{readNcVarsAndAttrs} extracts variables with long names as attributes.
\code{addTimeHoursFrom1950} adds time from the "hours since 1950"-variable of the NORWECOM data.
\code{getBiomassXY} converts to Cartesian coordinates for the biomass data.
\code{getSuperIndXY} converts to Cartesian coordinates for the superindividual data.
}
\keyword{internal}
