% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{reportAllPelfoss}
\alias{reportAllPelfoss}
\title{Function to generate all reports from a PELFOSS survey.}
\usage{
reportAllPelfoss(run, dir, copy = TRUE)
}
\arguments{
\item{run}{The name of the simulation run (which is three timings (-30, 0, +30 days), two directions (normal, reversed), and two resolutions (low, high), in total 12 individual runs).}

\item{dir}{The directory of the PELFOSS folder structure (see \code{\link{getPelfossSkeleton}}).}

\item{copy}{Logical: If TRUE copy the files form the individual runs to the reports directory.}
}
\description{
Function to generate all reports from a PELFOSS survey.
}
