% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{runAll}
\alias{runAll}
\title{Function for running all combinations of parameters of a survey stored in the spec, which is a data frame with one row per run.}
\usage{
runAll(spec, run, dir, surveyPar, seed = 1, nboot = 100, radius = 10,
  subset = "all", ...)
}
\arguments{
\item{spec}{A data frame of the specifications of res and year (these are linked in the current PELFOSS framework), dateshift and reversed.}

\item{run}{The name of the simulation run (which is three timings (-30, 0, +30 days), two directions (normal, reversed), and two resolutions (low, high), in total 12 individual runs).}

\item{dir}{The directory of the PELFOSS folder structure (see \code{\link{getPelfossSkeleton}}).}

\item{surveyPar}{A list of survey parameters as returned from \code{\link{setSurveyParametersPaper1}}.}

\item{seed}{A single integer, or a list of seeds used in the funciton, including the following seeds: ('transect') for drawing transects using surveyPlanner(), ('trawl') for drawing trawl stations along the transects with probability as a funciton of the NASC (see \code{probfact}), and ('bootstrap') for getting the final estimate using runBootstrap() (see \code{nboot}).}

\item{nboot}{Number of bootstrap replicates.}

\item{radius}{A vector of one or two elements giving the radius inside which to sample trawl stations, and the radius inside which to estimate the parameters a, b and L in the length-weight relationship. If only one element is given, the length-weight relationship is estimated from all adult super individuals.}

\item{subset}{Either "all" (the default) or an integer vector giving the subset of the simulations to run.}

\item{...}{Additional inputs overriding the defaults returned by pelfossDefaults().}
}
\description{
Function for running all combinations of parameters of a survey stored in the spec, which is a data frame with one row per run.
}
