% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{runAllPelfossPaper1}
\alias{runAllPelfossPaper1}
\title{Funciton for running all specified surveys of paper 1 of the PELFOSS project.}
\usage{
runAllPelfossPaper1(type = "Parallel", run = type,
  survey = "Herring_IESNS", dir = "~/Projects/PELFOSS/delphi",
  speedFact = 1, condPar = NULL, LcmMean = NULL, year = c(2010,
  2012), res = c(4, 10), dateshift = c(-30, 0, 30),
  reversed = c(FALSE, TRUE), seed = 1, nboot = 100, radius = 10,
  subset = "all", cores = 1, ...)
}
\arguments{
\item{type}{The survey design type (see \code{\link[Rstox]{surveyPlanner}}).}

\item{run}{The name of the simulation run (which is three timings (-30, 0, +30 days), two directions (normal, reversed), and two resolutions (low, high), in total 12 individual runs).}

\item{survey}{The name of the survey to run, one of "Herring_IESNS", "Herring_NASSHS", "Herring_NASSHS_fishery", "Mackerel_IESSNS", "Mackerel_IESSNS_sept" and "Mackerel_IESSNS_sept_fishery".}

\item{dir}{The directory of the PELFOSS folder structure (see \code{\link{getPelfossSkeleton}}).}

\item{speedFact}{A factor to speed the survey up or down by. Values larger than 1 speed up and values smaller than 1 slow down.}

\item{condPar}{A vector of two elements giving the parameters a and b in the length-weight relationship Wg = a * Lcm^b, where Wg is the weight in grams and Lcm is the length in cm. Typical values are e.g., a = 0.01 and b = 3. If not set, this will be estimated from the super individuals, possibly inside a radius given as the second elements of \code{radius}.}

\item{LcmMean}{The typical length of the fish, representing the mean acoustic backscatter which refects the square of fish length. This can typically be estimated as sqrt(mean(L^2)). If not set, this will be estimated from the super individuals, possibly inside a radius given as the second elements of \code{radius}.}

\item{year, res, dateshift, reversed}{These parameters are linked to the folder structure specifide by \code{\link{getPelfossSkeleton}}, and specifies the year/resolusion, which are linked in the PELFOSS folder structure, the shift in date, which is a vector of integer days to shift the survey by, and a logical stating whether to reverse the direction of the survey. See the default values.}

\item{seed}{A single integer, or a list of seeds used in the funciton, including the following seeds: ('transect') for drawing transects using surveyPlanner(), ('trawl') for drawing trawl stations along the transects with probability as a funciton of the NASC (see \code{probfact}), and ('bootstrap') for getting the final estimate using runBootstrap() (see \code{nboot}).}

\item{nboot}{Number of bootstrap replicates.}

\item{radius}{A vector of one or two elements giving the radius inside which to sample trawl stations, and the radius inside which to estimate the parameters a, b and L in the length-weight relationship. If only one element is given, the length-weight relationship is estimated from all adult super individuals.}

\item{subset}{Either "all" (the default) or an integer vector giving the subset of the simulations to run.}

\item{cores}{The number of cores to use for writing XML files and for the bootstrapping in the StoX project, given either as a named list such as list(biotic=1, acoustic=1, bootstrap=1), og a single numeric setting the cores for all.}

\item{...}{Additional inputs overriding the defaults returned by pelfossDefaults().}
}
\description{
Funciton for running all specified surveys of paper 1 of the PELFOSS project.
}
