% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{setSurveyParametersPaper1}
\alias{setSurveyParametersPaper1}
\title{Function for setting the parameters of each survey for paper 1 using the PELFOSS framework.}
\usage{
setSurveyParametersPaper1(
  run,
  dir,
  survey = "Herring_IESNS",
  type = "RectEnclZZ"
)
}
\arguments{
\item{run}{The name of the simulation run (which is three timings (-30, 0, +30 days), two directions (normal, reversed), and two resolutions (low, high), in total 12 individual runs).}

\item{dir}{The directory of the PELFOSS folder structure (see \code{\link{getPelfossSkeleton}}).}

\item{survey}{The name of the survey to run, one of "Herring_IESNS", "Herring_NASSHS", "Herring_NASSHS_fishery", "Mackerel_IESSNS", "Mackerel_IESSNS_sept" and "Mackerel_IESSNS_sept_fishery".}

\item{type}{The survey design type (see \code{\link[Rstox]{surveyPlanner}}).}
}
\description{
Function for setting the parameters of each survey for paper 1 using the PELFOSS framework.
}
