% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{simulateTrawl}
\alias{simulateTrawl}
\alias{drawTrawlStationInd}
\title{Simulate trawl stations given NORWECOM superindividuals.}
\usage{
simulateTrawl(superInd, fishStation, seed = 0, radius = 10, nn = NULL,
  N = 100, lengthunit = 2)

drawTrawlStationInd(nasc, nstations = 1, seed = 0, probfact = 0)
}
\arguments{
\item{x}{The input NetCDF4 file.}
}
\value{

}
\description{
Simulate trawl stations given NORWECOM superindividuals.
}
