% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{getSuperIndOfDay}
\alias{getSuperIndOfDay}
\title{Extract one day of valid superindividuals.}
\usage{
getSuperIndOfDay(superInd, day = NULL, maxValue = 1e+10,
  NAsByFirst = FALSE)
}
\arguments{
\item{x}{The input NetCDF4 file.}
}
\value{

}
\description{
Extract one day of valid superindividuals.
}
\keyword{internal}
