% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{getSuperindOfDay}
\alias{getSuperindOfDay}
\alias{getBiomassOfSurvey}
\alias{getTotalBiomass}
\title{Extract dayly subsets of biomass and superind.}
\usage{
getSuperindOfDay(superind, superindDay = NULL, superindFilter = NULL,
  stratumPolygons = NULL, transectsXyOfDay = NULL,
  insideRange = NULL)

getBiomassOfSurvey(x, daysOfSurvey = NULL, fresh = FALSE)

getTotalBiomass(superind, stratumPolygons, strataInd = NULL,
  days = NULL, type = c("biomass", "length"), superindFilter = NULL,
  union = FALSE, transectsXyOfDay = NULL, insideRange = NULL)
}
\arguments{
\item{superindFilter}{A function such as superindFilter <- function(x){x$Long < 20 & x$Lat < 70}, taking the superind data as input and returning a logical vector discarding certain longitude/latitude combinations.}

\item{stratumPolygons}{A named list of stratum polygons as data frames with longitude in the first column and latitude in the second.}

\item{daysOfSurvey}{A vector of the days of the survey over which to average the biomass.}

\item{strataInd}{A vector of indices of the strata to compute the total biomass over.}

\item{days}{A vector of indices of the days to compute the total biomass for.}

\item{type}{A string specifying the type of calculation (first element used), one of "biomass" and "length", returning total biomass and the length that would result in unbiased estimated total biomass if bioass was converted to NASC and aggregated throughout the survey region.}

\item{biomass, superind}{The biomass and superind data read from \code{readNcVarsAndAttrs} (used in \code{readNorwecomBiomass} and \code{readNorwecomSuperind}).}

\item{day}{The day on which to extract the superind data.}

\item{NAsByFirst}{Logical: If TRUE, identify NAs by the first day of the superind data.}

\item{midDayOfSurvey}{The mid day of the survey ar which to extract the superind data.}
}
\description{
\code{getSuperindOfDay} Gets the superind data of one day. \cr \cr
\code{getBiomassOfSurvey} Gets the average biomass data of all days of a survey. \cr \cr
\code{getTotalBiomass} Gets the total biomass/'fish length which would produce an unbiased biomass via acoustic backscattering'. Uses the superind data. \cr \cr
}
\keyword{internal}
