% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{pelfossDefaults}
\alias{pelfossDefaults}
\alias{getPelfossPaths}
\alias{createPelfossSkeleton}
\alias{fitLengthWeightSuperindOld}
\alias{fitLengthWeightSuperind}
\alias{fitLengthWeight}
\alias{plotLengthWeight}
\title{Utility functions of the pelfoss package.}
\usage{
pelfossDefaults()

getPelfossPaths(dir, survey = "Herring_IESNS", year = 2010, res = 4,
  reversed = FALSE, dateshift = 0, seed = 0)

createPelfossSkeleton(dir)

fitLengthWeightSuperindOld(day, superind, superindFilter = NULL, b = NULL)

fitLengthWeightSuperind(day, superind, count = "inumb",
  superindFilter = NULL, a = NULL, b = NULL)

fitLengthWeight(L, W = NULL, a = NULL, b = NULL, count = NULL,
  filter = NULL)

plotLengthWeight(x, cols = c(1, 2), useJitter = FALSE, ...)
}
\arguments{
\item{dir}{The root directory of the pelfoss folder structure.}

\item{survey}{The survey name.}

\item{year}{The year.}

\item{res}{The resoludion.}

\item{reversed}{Logical: If TRUE reverse the direction of the survey as generated by \code{\link{surveyPlanner}}.}

\item{dateshift}{The number of days to shift the timing of the survey by.}

\item{seed}{The seed used in \code{\link{surveyPlanner}}.}
}
\description{
\code{pelfossDefaults} Defines default values which can be overridden by the \code{...} option in \code{runPelfoss}. \cr \cr
\code{getPelfossPaths} Given a root directory of a pelfoss folder structure (see ?pelfoss) and a survey name, this function returns the paths to biomass and superind NetCDF4-files, stratum polygons, fishery data files (any data format) of the species as interpreted from the survey name. . \cr \cr
}
\keyword{internal}
