% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelfoss.R
\name{runPelfoss}
\alias{runPelfoss}
\alias{pelfossDefaults}
\alias{getNorwecomPaths}
\alias{getAllTSB}
\title{Run the PELFOSS framework.}
\usage{
runPelfoss(dir, surveyPar, year = 2010, res = 4, dateshift = 0,
  reversed = FALSE, seed = 0, projectName = "NORWECOM_Rstox",
  nboot = 10, xsd = list(acoustic = "1", biotic = "1.4"), format = "png",
  ...)

pelfossDefaults()

getNorwecomPaths(dir, survey = "Herring_IESNS", year = 2010, res = 4,
  reversed = FALSE, dateshift = 0, seed = 0)

getAllTSB(dir, outfile = NULL)
}
\arguments{
\item{dir}{The directory holding the NORWECOM files, the stratum files and the output xml files from the function (temporary stored for insertion into the StoX project).}

\item{year, res}{The year and resolution (in km) to run. The NORWECOM files are located in the following folder structure: species / resolution / year, where the resolution folders has names such as "res_4km".}

\item{reversed}{Logical: If TRUE run the survey in the opposite direciton.}

\item{seed}{A single integer, or a list of seeds used in the funciton, including the following seeds: ('transect') for drawing transects using surveyPlanner(), ('trawl') for drawing trawl stations along the transects with probability as a funciton of the NASC (see \code{probfact}), and ('bootstrap') for getting the final estimate using runBootstrap() (see \code{nboot}).}

\item{projectName}{The name or full path of the StoX project of the simulated survey.}

\item{nboot}{Number of bootstrap replicates.}

\item{xsd}{A named list of xsd versions of the acoustic and biotic file format.}

\item{...}{Additional inputs overriding the defaults returned by pelfossDefaults().}

\item{survey}{The name of the survey to simulate, one of "Herring_IESNS" (International ecosystem survey in the Nordic Seas), "Herring_NASSHS" (Norwegian acoustic spring spawning herring survey) or "Mackerel_IESSNS" (The International Ecosystem Summer Survey in the Nordic Seas)}

\item{type}{The type of survey. See \code{\link{surveyPlanner}}}
}
\value{

}
\description{
This function runs the entire PELFOSS framework from the NORWECOM model output files (biomass in a grid and superindividuals), through surveyPlanner() and StoX to the estimated total stock biomass compared to the corresponding theoretical values:
}
\keyword{internal}
